Use of Sockets to watch for Misconduct updates
A new socket route was created to facilitate students listening to
reported misconducts. On start up the user id is set in the socket
and a watch stream is set up on the ExamAllocation model, emitting to
the listener when their allocation is updated. It emits the length of the
misconduct array.
id_numer was also added to returned user information from session
variable